pyproject.toml
[tool.poetry]
name = "user-profile-service"
version = "0.1.0"
description = ""
authors = ["Boris Averin <89068157313@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = ">=0.115.12,<0.116.0"
sqlalchemy = ">=2.0.40,<3.0.0"
alembic = ">=1.15.2,<2.0.0"
pydantic-settings = ">=2.8.1,<3.0.0"
uvicorn = ">=0.34.0,<0.35.0"
asyncpg = ">=0.30.0,<0.31.0"
greenlet = ">=3.1.1,<4.0.0"
geoalchemy2 = ">=0.17.1,<0.18.0"
pydantic-extra-types = ">=2.10.3,<3.0.0"
httpx = ">=0.28.1,<0.29.0"
python-jose = ">=3.4.0,<4.0.0"
asyncio-redis = ">=0.16.0,<0.17.0"
redis = ">=5.2.1,<6.0.0"
python-dateutil = "^2.9.0.post0"


[tool.poetry.group.dev.dependencies]
ruff = ">=0.12.2,<0.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
fix = true

exclude = [
    ".git",
    ".venv",
    "migrations",
    "__pycache__"
]

# выбор наборов правил
select = [
    "E", # pycodestyle errors
    "F", # pyflakes
    "I", # isort
    "UP", # pyupgrade
    "N", # pep8-naming
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "RUF", # Ruff-specific rules
]

line-length = 88

ignore = [
    "B008",
    "RUF001",
    "RUF002",
    "N812"
]